@@ -1,16 +1,21 @@
-const { awscdk } = require('projen');
+import { CdklabsConstructLibrary } from 'cdklabs-projen-project-types';
 
-const project = new awscdk.AwsCdkConstructLibrary({
+const project = new CdklabsConstructLibrary({
   author: 'Amazon Web Services, Inc.',
+  projenrcTs: true,
+  private: false,
+  enablePRAutoMerge: true,
   authorAddress: 'aws-cdk-team@amazon.com',
   cdkVersion: '2.1.0',
   defaultReleaseBranch: 'main',
   name: 'cdk-drift-monitor',
-  minNodeVersion: '14.18.0',
+
+  workflowNodeVersion: '16.x',
+  minNodeVersion: '16.0.0',
+
   description: 'Monitors for CloudFormation stack drifts',
   repositoryUrl: 'https://github.com/cdklabs/cdk-drift-monitor',
   packageName: 'cdk-drift-monitor',
-  projenUpgradeSecret: 'PROJEN_GITHUB_TOKEN',
   deps: [
     'aws-cdk-lib',
     'constructs',
@@ -25,11 +30,16 @@ const project = new awscdk.AwsCdkConstructLibrary({
     'jest-each',
     'aws-cdk-lib',
     'constructs',
+    'cdklabs-projen-project-types',
   ],
   gitignore: [
     'src/**/*.js',
   ],
   releaseToNpm: true,
+
+  // do not release to other jsii languages because thats the way it was
+  // when migrating to cdklabs-projen-project-types.
+  jsiiTargetLanguages: [],
   autoApproveOptions: {
     allowedUsernames: ['cdklabs-automation'],
     secret: 'GITHUB_TOKEN',
