@@ -1,4 +1,4 @@
-// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+// ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
 import * as path from 'path';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 import { Construct } from 'constructs';
